requirements.txt
# requirements.txt
fastapi>=0.95.0,<0.96.0
uvicorn>=0.22.0,<0.23.0
SQLAlchemy>=1.4.47,<1.5.0
pydantic>=1.10.18,<2.0.0
click>=8.1.3,<9.0.0
rich>=13.5.2,<14.0.0
PyQt5>=5.15.9,<6.0.0
gspread>=5.7.2,<6.0.0
oauth2client>=4.1.3,<5.0.0
pandas>=2.2.3,<3.0.0
fuzzywuzzy>=0.18.0,<0.19.0
python-dotenv>=1.0.0,<2.0.0

